feat: create ChainsDropdown component
